extract markdown rendering and highlighting out
