+++ Cargo.toml
@@ -22,7 +22,7 @@ name = "deno_doc"
 
 [[example]]
 name = "ddoc"
-required-features = ["html"]
+required-features = ["html", "comrak"]
 
 [dependencies]
 anyhow = "1.0.86"
@@ -37,32 +37,12 @@ regex = "1.10.6"
 serde.workspace = true
 serde_json = { version = "1.0.122", features = ["preserve_order"] }
 termcolor = "1.4.1"
+itoa = "1.0.11"
 
 html-escape = { version = "0.2.13", optional = true }
-comrak = { version = "0.28.0", optional = true, default-features = false }
-handlebars = { version = "6.1", optional = true, features = ["string_helpers"] }
-syntect = { version = "5.2.0", optional = true, default-features = false, features = [
-  "parsing",
-  "default-syntaxes",
-  "default-themes",
-  "html",
-  "dump-load",
-  "regex-onig",
-] }
+handlebars = { version = "6.1", features = ["string_helpers"], optional = true }
+comrak = { version = "0.29.0", optional = true, default-features = false }
 ammonia = { version = "4.0.0", optional = true }
-
-tree-sitter-highlight = { version = "0.22.6", optional = true }
-tree-sitter-javascript = { version = "0.21.4", optional = true }
-tree-sitter-typescript = { version = "0.21.2", optional = true }
-tree-sitter-json = { version = "0.21.0", optional = true }
-tree-sitter-regex = { version = "0.21.0", optional = true }
-tree-sitter-css = { version = "0.21.0", optional = true }
-tree-sitter-md = { version = "0.2.3", optional = true }
-tree-sitter-rust = { version = "0.21.2", optional = true }
-tree-sitter-html = { version = "0.20.3", optional = true }
-tree-sitter-bash = { version = "0.21.0", optional = true }
-tree-sitter-xml = { version = "0.6.4", optional = true }
-itoa = "1.0.11"
 
 [dev-dependencies]
 anyhow = { version = "1.0.86" }
@@ -75,22 +55,10 @@ pretty_assertions = "1.4.0"
 insta = { version = "1.39.0", features = ["json"] }
 
 [features]
-default = ["html", "rust", "tree-sitter"]
+default = ["rust", "html"]
 rust = []
-html = ["html-escape", "comrak", "handlebars", "ammonia"]
-tree-sitter = [
-  "tree-sitter-highlight",
-  "tree-sitter-javascript",
-  "tree-sitter-typescript",
-  "tree-sitter-json",
-  "tree-sitter-regex",
-  "tree-sitter-css",
-  "tree-sitter-md",
-  "tree-sitter-rust",
-  "tree-sitter-html",
-  "tree-sitter-bash",
-  "tree-sitter-xml",
-]
+html = ["html-escape", "handlebars"]
+comrak = ["dep:comrak", "ammonia"]
 
 [[test]]
 name = "specs"
