+++ src/html/jsdoc.rs
@@ -1,20 +1,11 @@
 use super::render_context::RenderContext;
 use super::util::*;
-use crate::html::comrak_adapters::URLRewriter;
 use crate::html::ShortPath;
 use crate::js_doc::JsDoc;
 use crate::js_doc::JsDocTag;
 use crate::DocNodeKind;
-use comrak::nodes::Ast;
-use comrak::nodes::AstNode;
-use comrak::nodes::NodeHtmlBlock;
-use comrak::nodes::NodeValue;
-use comrak::Arena;
 use serde::Serialize;
 use std::borrow::Cow;
-use std::cell::RefCell;
-use std::io::BufWriter;
-use std::io::Write;
 
 lazy_static! {
   static ref JSDOC_LINK_RE: regex::Regex = regex::Regex::new(
@@ -25,76 +16,6 @@ lazy_static! {
     regex::Regex::new(r"(^\.{0,2}\/)|(^[A-Za-z]+:\S)").unwrap();
   static ref MODULE_LINK_RE: regex::Regex =
     regex::Regex::new(r"^\[(\S+)\](?:\.(\S+)|\s|)$").unwrap();
-}
-
-lazy_static! {
-  static ref AMMONIA: ammonia::Builder<'static> = {
-    let mut ammonia_builder = ammonia::Builder::default();
-
-    ammonia_builder
-      .add_tags(["video", "button", "svg", "path", "rect"])
-      .add_generic_attributes(["id", "align"])
-      .add_tag_attributes("button", ["data-copy"])
-      .add_tag_attributes(
-        "svg",
-        [
-          "width",
-          "height",
-          "viewBox",
-          "fill",
-          "xmlns",
-          "stroke",
-          "stroke-width",
-          "stroke-linecap",
-          "stroke-linejoin",
-        ],
-      )
-      .add_tag_attributes(
-        "path",
-        [
-          "d",
-          "fill",
-          "fill-rule",
-          "clip-rule",
-          "stroke",
-          "stroke-width",
-          "stroke-linecap",
-          "stroke-linejoin",
-        ],
-      )
-      .add_tag_attributes("rect", ["x", "y", "width", "height", "fill"])
-      .add_tag_attributes("video", ["src", "controls"])
-      .add_allowed_classes("pre", ["highlight"])
-      .add_allowed_classes("button", ["context_button"])
-      .add_allowed_classes(
-        "div",
-        [
-          "alert",
-          "alert-note",
-          "alert-tip",
-          "alert-important",
-          "alert-warning",
-          "alert-caution",
-        ],
-      )
-      .link_rel(Some("nofollow"))
-      .url_relative(ammonia::UrlRelative::Custom(Box::new(
-        AmmoniaRelativeUrlEvaluator(),
-      )));
-
-    #[cfg(feature = "syntect")]
-    ammonia_builder.add_tag_attributes("span", ["style"]);
-
-    #[cfg(feature = "tree-sitter")]
-    ammonia_builder.add_allowed_classes("span", super::tree_sitter::CLASSES);
-
-    ammonia_builder
-  };
-}
-
-thread_local! {
-  static CURRENT_FILE: RefCell<Option<Option<ShortPath>>> = const { RefCell::new(None) };
-  static URL_REWRITER: RefCell<Option<Option<URLRewriter>>> = const { RefCell::new(None) };
 }
 
 fn parse_links<'a>(md: &'a str, ctx: &RenderContext) -> Cow<'a, str> {
@@ -219,323 +140,75 @@ fn split_markdown_title(md: &str) -> (Option<&str>, Option<&str>) {
   }
 }
 
-struct AmmoniaRelativeUrlEvaluator();
-
-impl<'b> ammonia::UrlRelativeEvaluate<'b> for AmmoniaRelativeUrlEvaluator {
-  fn evaluate<'a>(&self, url: &'a str) -> Option<Cow<'a, str>> {
-    URL_REWRITER.with(|url_rewriter| {
-      if let Some(url_rewriter) = url_rewriter.borrow().as_ref().unwrap() {
-        CURRENT_FILE.with(|current_file| {
-          Some(
-            url_rewriter(current_file.borrow().as_ref().unwrap().as_ref(), url)
-              .into(),
-          )
-        })
-      } else {
-        Some(Cow::Borrowed(url))
-      }
-    })
-  }
-}
-
-enum Alert {
-  Note,
-  Tip,
-  Important,
-  Warning,
-  Caution,
-}
-
-fn match_node_value<'a>(
-  arena: &'a Arena<AstNode<'a>>,
-  node: &'a AstNode<'a>,
-  options: &comrak::Options,
-  plugins: &comrak::Plugins,
-) {
-  match &node.data.borrow().value {
-    NodeValue::BlockQuote => {
-      if let Some(paragraph_child) = node.first_child() {
-        if paragraph_child.data.borrow().value == NodeValue::Paragraph {
-          let alert = paragraph_child.first_child().and_then(|text_child| {
-            if let NodeValue::Text(text) = &text_child.data.borrow().value {
-              match text
-                .split_once(' ')
-                .map_or((text.as_str(), None), |(kind, title)| {
-                  (kind, Some(title))
-                }) {
-                ("[!NOTE]", title) => {
-                  Some((Alert::Note, title.unwrap_or("Note").to_string()))
-                }
-                ("[!TIP]", title) => {
-                  Some((Alert::Tip, title.unwrap_or("Tip").to_string()))
-                }
-                ("[!IMPORTANT]", title) => Some((
-                  Alert::Important,
-                  title.unwrap_or("Important").to_string(),
-                )),
-                ("[!WARNING]", title) => {
-                  Some((Alert::Warning, title.unwrap_or("Warning").to_string()))
-                }
-                ("[!CAUTION]", title) => {
-                  Some((Alert::Caution, title.unwrap_or("Caution").to_string()))
-                }
-                _ => None,
-              }
-            } else {
-              None
-            }
-          });
-
-          if let Some((alert, title)) = alert {
-            let start_col = node.data.borrow().sourcepos.start;
-
-            let document = arena.alloc(AstNode::new(RefCell::new(Ast::new(
-              NodeValue::Document,
-              start_col,
-            ))));
-
-            let node_without_alert = arena.alloc(AstNode::new(RefCell::new(
-              Ast::new(NodeValue::Paragraph, start_col),
-            )));
-
-            for child_node in paragraph_child.children().skip(1) {
-              node_without_alert.append(child_node);
-            }
-            for child_node in node.children().skip(1) {
-              node_without_alert.append(child_node);
-            }
-
-            document.append(node_without_alert);
-
-            let html = render_node(document, options, plugins);
-
-            let alert_title = match alert {
-              Alert::Note => format!(
-                "{}{title}",
-                include_str!("./templates/icons/info-circle.svg")
-              ),
-              Alert::Tip => {
-                format!("{}{title}", include_str!("./templates/icons/bulb.svg"))
-              }
-              Alert::Important => format!(
-                "{}{title}",
-                include_str!("./templates/icons/warning-message.svg")
-              ),
-              Alert::Warning => format!(
-                "{}{title}",
-                include_str!("./templates/icons/warning-triangle.svg")
-              ),
-              Alert::Caution => format!(
-                "{}{title}",
-                include_str!("./templates/icons/warning-octagon.svg")
-              ),
-            };
-
-            let html = format!(
-              r#"<div class="alert alert-{}"><div>{alert_title}</div><div>{html}</div></div>"#,
-              match alert {
-                Alert::Note => "note",
-                Alert::Tip => "tip",
-                Alert::Important => "important",
-                Alert::Warning => "warning",
-                Alert::Caution => "caution",
-              }
-            );
-
-            let alert_node = arena.alloc(AstNode::new(RefCell::new(Ast::new(
-              NodeValue::HtmlBlock(NodeHtmlBlock {
-                block_type: 6,
-                literal: html,
-              }),
-              start_col,
-            ))));
-            node.insert_before(alert_node);
-            node.detach();
-          }
-        }
-      }
-    }
-    NodeValue::Link(link) => {
-      if link.url.ends_with(".mov") || link.url.ends_with(".mp4") {
-        let start_col = node.data.borrow().sourcepos.start;
-
-        let html = format!(r#"<video src="{}" controls></video>"#, link.url);
-
-        let alert_node = arena.alloc(AstNode::new(RefCell::new(Ast::new(
-          NodeValue::HtmlBlock(NodeHtmlBlock {
-            block_type: 6,
-            literal: html,
-          }),
-          start_col,
-        ))));
-        node.insert_before(alert_node);
-        node.detach();
-      }
-    }
-    _ => {}
-  }
-}
-
-fn walk_node<'a>(
-  arena: &'a Arena<AstNode<'a>>,
-  node: &'a AstNode<'a>,
-  options: &comrak::Options,
-  plugins: &comrak::Plugins,
-) {
-  for child in node.children() {
-    match_node_value(arena, child, options, plugins);
-    walk_node(arena, child, options, plugins);
-  }
-}
-
-fn walk_node_title<'a>(node: &'a AstNode<'a>) {
-  for child in node.children() {
-    if matches!(
-      child.data.borrow().value,
-      NodeValue::Document
-        | NodeValue::Paragraph
-        | NodeValue::Heading(_)
-        | NodeValue::Text(_)
-        | NodeValue::Code(_)
-        | NodeValue::HtmlInline(_)
-        | NodeValue::Emph
-        | NodeValue::Strong
-        | NodeValue::Strikethrough
-        | NodeValue::Superscript
-        | NodeValue::Link(_)
-        | NodeValue::Math(_)
-        | NodeValue::Escaped
-        | NodeValue::WikiLink(_)
-        | NodeValue::Underline
-        | NodeValue::SoftBreak
-    ) {
-      walk_node_title(child);
-    } else {
-      // delete the node
-      child.detach();
-    }
-  }
-}
-
-fn render_node<'a>(
-  node: &'a AstNode<'a>,
-  options: &comrak::Options,
-  plugins: &comrak::Plugins,
-) -> String {
-  let mut bw = BufWriter::new(Vec::new());
-  comrak::format_html_with_plugins(node, options, &mut bw, plugins).unwrap();
-  String::from_utf8(bw.into_inner().unwrap()).unwrap()
-}
-
 pub struct MarkdownToHTMLOptions {
   pub title_only: bool,
   pub no_toc: bool,
 }
 
-pub fn strip(render_ctx: &RenderContext, md: &str) -> String {
-  let mut options = comrak::Options::default();
-  options.extension.autolink = true;
-  options.extension.description_lists = true;
-  options.extension.strikethrough = true;
-  options.extension.superscript = true;
-  options.extension.table = true;
-  options.extension.tagfilter = true;
-  options.extension.tasklist = true;
-  options.render.escape = true;
+pub type MarkdownStripper = std::rc::Rc<dyn (Fn(&str) -> String)>;
 
+pub fn strip(render_ctx: &RenderContext, md: &str) -> String {
   let md = parse_links(md, render_ctx);
 
-  let arena = Arena::new();
-  let root = comrak::parse_document(&arena, &md, &options);
+  (render_ctx.ctx.markdown_stripper)(&md)
+}
 
   walk_node(&arena, root, &options, &Default::default());
 
-  fn collect_text<'a>(node: &'a AstNode<'a>, output: &mut BufWriter<Vec<u8>>) {
-    match node.data.borrow().value {
-      NodeValue::Text(ref literal)
-      | NodeValue::Code(comrak::nodes::NodeCode { ref literal, .. }) => {
-        output.write_all(literal.as_bytes()).unwrap();
-      }
-      NodeValue::LineBreak | NodeValue::SoftBreak => {
-        output.write_all(&[b' ']).unwrap()
-      }
-      _ => {
-        for n in node.children() {
-          collect_text(n, output);
-        }
-      }
-    }
-  }
+pub type Anchorizer =
+  std::sync::Arc<dyn Fn(String, u8) -> String + Send + Sync>;
 
-  let mut bw = BufWriter::new(Vec::new());
-  collect_text(root, &mut bw);
-  String::from_utf8(bw.into_inner().unwrap()).unwrap()
-}
+pub type MarkdownRenderer = std::rc::Rc<
+  dyn (Fn(&str, bool, Option<ShortPath>, Anchorizer) -> Option<String>),
+>;
 
 pub fn markdown_to_html(
   render_ctx: &RenderContext,
   md: &str,
   render_options: MarkdownToHTMLOptions,
 ) -> Option<String> {
-  // TODO(bartlomieju): this should be initialized only once
-  let mut options = comrak::Options::default();
-  options.extension.autolink = true;
-  options.extension.description_lists = true;
-  options.extension.strikethrough = true;
-  options.extension.superscript = true;
-  options.extension.table = true;
-  options.extension.tagfilter = true;
-  options.extension.tasklist = true;
-  options.render.unsafe_ = true; // its fine because we run ammonia afterwards
+  let toc = render_ctx.toc.clone();
 
-  let mut plugins = comrak::Plugins::default();
+  let anchorizer = move |content: String, level: u8| {
+    let mut anchorizer = toc.anchorizer.lock().unwrap();
+    let offset = toc.offset.lock().unwrap();
 
-  if !render_options.title_only {
-    plugins.render.codefence_syntax_highlighter =
-      Some(&render_ctx.ctx.highlight_adapter);
+    let anchor = anchorizer.anchorize(&content);
+
     if !render_options.no_toc {
-      plugins.render.heading_adapter = Some(&render_ctx.toc);
+      let mut toc = toc.toc.lock().unwrap();
+      toc.push(crate::html::render_context::ToCEntry {
+        level: level + *offset,
+        content,
+        anchor: anchor.clone(),
+      });
     }
-  }
+
+    anchor
+  };
+
+  #[cfg(not(target_arch = "wasm32"))]
+  let anchorizer = std::sync::Arc::new(anchorizer);
+
+  #[cfg(target_arch = "wasm32")]
+  let anchorizer = wasm_bindgen::prelude::Closure::wrap(
+    Box::new(anchorizer) as Box<dyn Fn(String, u8) -> String>
+  );
+  #[cfg(target_arch = "wasm32")]
+  let anchorizer = &wasm_bindgen::JsCast::unchecked_ref::<js_sys::Function>(
+    anchorizer.as_ref(),
+  );
 
   let md = parse_links(md, render_ctx);
 
-  let class_name = if render_options.title_only {
-    "markdown_summary"
-  } else {
-    "markdown"
-  };
+  let file = render_ctx.get_current_resolve().get_file().cloned();
 
-  let html = {
-    let arena = Arena::new();
-    let root = comrak::parse_document(&arena, &md, &options);
-
-    if render_options.title_only {
-      walk_node_title(root);
-
-      if let Some(child) = root.first_child() {
-        render_node(child, &options, &plugins)
-      } else {
-        return None;
-      }
-    } else {
-      walk_node(&arena, root, &options, &plugins);
-      render_node(root, &options, &plugins)
-    }
-  };
-
-  CURRENT_FILE.set(Some(render_ctx.get_current_resolve().get_file().cloned()));
-  URL_REWRITER.set(Some(render_ctx.ctx.url_rewriter.clone()));
-
-  let html = Some(format!(
-    r#"<div class="{class_name}">{}</div>"#,
-    AMMONIA.clean(&html)
-  ));
-
-  CURRENT_FILE.set(None);
-  URL_REWRITER.set(None);
-
-  html
+  (render_ctx.ctx.markdown_renderer)(
+    &md,
+    render_options.title_only,
+    file,
+    anchorizer,
+  )
 }
 
 pub(crate) fn render_markdown(
@@ -718,21 +391,26 @@ impl ModuleDocCtx {
 mod test {
   use crate::html::href_path_resolve;
   use crate::html::jsdoc::parse_links;
+  use crate::html::DocNodeWithContext;
   use crate::html::GenerateCtx;
   use crate::html::GenerateOptions;
   use crate::html::HrefResolver;
+  use crate::html::UsageComposer;
+  use crate::html::UsageComposerEntry;
   use crate::DocNode;
   use crate::Location;
   use deno_ast::ModuleSpecifier;
   use indexmap::IndexMap;
+  use std::rc::Rc;
 
+  use crate::html::usage::UsageToMd;
   use crate::html::RenderContext;
   use crate::html::UrlResolveKind;
   use crate::interface::InterfaceDef;
   use crate::js_doc::JsDoc;
   use crate::node::DeclarationKind;
 
-  struct EmptyResolver {}
+  struct EmptyResolver;
 
   impl HrefResolver for EmptyResolver {
     fn resolve_path(
@@ -755,12 +433,6 @@ mod test {
       None
     }
 
-    fn resolve_usage(&self, current_resolve: UrlResolveKind) -> Option<String> {
-      current_resolve
-        .get_file()
-        .map(|current_file| current_file.display_name().to_string())
-    }
-
     fn resolve_source(&self, _location: &Location) -> Option<String> {
       None
     }
@@ -774,19 +446,47 @@ mod test {
     }
   }
 
+  impl UsageComposer for EmptyResolver {
+    fn is_single_mode(&self) -> bool {
+      true
+    }
+
+    fn compose(
+      &self,
+      nodes: &[DocNodeWithContext],
+      current_resolve: UrlResolveKind,
+      usage_to_md: UsageToMd,
+    ) -> IndexMap<UsageComposerEntry, String> {
+      current_resolve
+        .get_file()
+        .map(|current_file| {
+          IndexMap::from([(
+            UsageComposerEntry {
+              name: "".to_string(),
+              icon: None,
+            },
+            usage_to_md(nodes, current_file.display_name()),
+          )])
+        })
+        .unwrap_or_default()
+    }
+  }
+
   #[test]
   fn parse_links_test() {
     let ctx = GenerateCtx::new(
       GenerateOptions {
         package_name: None,
         main_entrypoint: None,
-        href_resolver: std::rc::Rc::new(EmptyResolver {}),
-        usage_composer: None,
+        href_resolver: Rc::new(EmptyResolver),
+        usage_composer: Rc::new(EmptyResolver),
         rewrite_map: None,
         category_docs: None,
         disable_search: false,
         symbol_redirect_map: None,
         default_symbol_map: None,
+        markdown_renderer: crate::html::comrak::create_renderer(None, None),
+        markdown_stripper: Rc::new(crate::html::comrak::strip),
       },
       Default::default(),
       Default::default(),
@@ -936,13 +636,15 @@ mod test {
       GenerateOptions {
         package_name: None,
         main_entrypoint: None,
-        href_resolver: std::rc::Rc::new(EmptyResolver {}),
-        usage_composer: None,
+        href_resolver: Rc::new(EmptyResolver),
+        usage_composer: Rc::new(EmptyResolver),
         rewrite_map: None,
         category_docs: None,
         disable_search: false,
         symbol_redirect_map: None,
         default_symbol_map: None,
+        markdown_renderer: crate::html::comrak::create_renderer(None, None),
+        markdown_stripper: Rc::new(crate::html::comrak::strip),
       },
       Default::default(),
       Default::default(),
