+++ src/html/mod.rs
@@ -11,7 +11,6 @@ use std::collections::HashMap;
 use std::path::PathBuf;
 use std::rc::Rc;
 
-pub mod comrak_adapters;
 pub mod jsdoc;
 pub mod pages;
 mod parameters;
@@ -19,11 +18,12 @@ pub mod partition;
 mod render_context;
 mod search;
 mod symbols;
-#[cfg(feature = "tree-sitter")]
-pub mod tree_sitter;
 mod types;
 mod usage;
 pub mod util;
+
+#[cfg(feature = "comrak")]
+pub mod comrak;
 
 use crate::html::pages::SymbolPage;
 use crate::js_doc::JsDocTag;
@@ -33,7 +33,9 @@ pub use search::generate_search_index;
 pub use symbols::namespace;
 pub use symbols::SymbolContentCtx;
 pub use symbols::SymbolGroupCtx;
-pub use usage::usage_to_md;
+pub use usage::UsageComposer;
+pub use usage::UsageComposerEntry;
+pub use usage::UsageToMd;
 pub use util::compute_namespaced_symbols;
 pub use util::href_path_resolve;
 pub use util::qualify_drilldown_name;
@@ -45,6 +47,8 @@ pub use util::ToCCtx;
 pub use util::TopSymbolCtx;
 pub use util::TopSymbolsCtx;
 pub use util::UrlResolveKind;
+
+pub const COPY_BUTTON: &str = include_str!("./templates/icons/copy.svg");
 
 pub const STYLESHEET: &str = include_str!("./templates/styles.gen.css");
 pub const STYLESHEET_FILENAME: &str = "styles.css";
@@ -203,10 +207,7 @@ fn setup_hbs() -> Result<Handlebars<'static>, anyhow::Error> {
     "icons/arrow",
     include_str!("./templates/icons/arrow.svg"),
   )?;
-  reg.register_template_string(
-    "icons/copy",
-    include_str!("./templates/icons/copy.svg"),
-  )?;
+  reg.register_template_string("icons/copy", COPY_BUTTON)?;
   reg.register_template_string(
     "icons/link",
     include_str!("./templates/icons/link.svg"),
@@ -227,20 +228,6 @@ lazy_static! {
   pub static ref HANDLEBARS: Handlebars<'static> = setup_hbs().unwrap();
 }
 
-pub type UsageComposer = Rc<
-  dyn Fn(
-    &RenderContext,
-    &[DocNodeWithContext],
-    String,
-  ) -> IndexMap<UsageComposerEntry, String>,
->;
-
-#[derive(Eq, PartialEq, Hash)]
-pub struct UsageComposerEntry {
-  pub name: String,
-  pub icon: Option<std::borrow::Cow<'static, str>>,
-}
-
 #[derive(Clone)]
 pub struct GenerateOptions {
   /// The name that is shown is the top-left corner, eg. "deno_std".
@@ -250,12 +237,14 @@ pub struct GenerateOptions {
   /// default to that file.
   pub main_entrypoint: Option<ModuleSpecifier>,
   pub href_resolver: Rc<dyn HrefResolver>,
-  pub usage_composer: Option<UsageComposer>,
+  pub usage_composer: Rc<dyn UsageComposer>,
   pub rewrite_map: Option<IndexMap<ModuleSpecifier, String>>,
   pub category_docs: Option<IndexMap<String, Option<String>>>,
   pub disable_search: bool,
   pub symbol_redirect_map: Option<IndexMap<String, IndexMap<String, String>>>,
   pub default_symbol_map: Option<IndexMap<String, String>>,
+  pub markdown_renderer: jsdoc::MarkdownRenderer,
+  pub markdown_stripper: jsdoc::MarkdownStripper,
 }
 
 #[non_exhaustive]
@@ -263,10 +252,8 @@ pub struct GenerateCtx {
   pub package_name: Option<String>,
   pub common_ancestor: Option<PathBuf>,
   pub doc_nodes: IndexMap<Rc<ShortPath>, Vec<DocNodeWithContext>>,
-  pub highlight_adapter: comrak_adapters::HighlightAdapter,
-  pub url_rewriter: Option<comrak_adapters::URLRewriter>,
   pub href_resolver: Rc<dyn HrefResolver>,
-  pub usage_composer: Option<UsageComposer>,
+  pub usage_composer: Rc<dyn UsageComposer>,
   pub rewrite_map: Option<IndexMap<ModuleSpecifier, String>>,
   pub main_entrypoint: Option<Rc<ShortPath>>,
   pub file_mode: FileMode,
@@ -274,6 +261,8 @@ pub struct GenerateCtx {
   pub disable_search: bool,
   pub symbol_redirect_map: Option<IndexMap<String, IndexMap<String, String>>>,
   pub default_symbol_map: Option<IndexMap<String, String>>,
+  pub markdown_renderer: jsdoc::MarkdownRenderer,
+  pub markdown_stripper: jsdoc::MarkdownStripper,
 }
 
 impl GenerateCtx {
@@ -368,8 +357,6 @@ impl GenerateCtx {
       package_name: options.package_name,
       common_ancestor,
       doc_nodes,
-      highlight_adapter: setup_highlighter(false),
-      url_rewriter: None,
       href_resolver: options.href_resolver,
       usage_composer: options.usage_composer,
       rewrite_map: options.rewrite_map,
@@ -379,6 +366,8 @@ impl GenerateCtx {
       disable_search: options.disable_search,
       symbol_redirect_map: options.symbol_redirect_map,
       default_symbol_map: options.default_symbol_map,
+      markdown_renderer: options.markdown_renderer,
+      markdown_stripper: options.markdown_stripper,
     })
   }
 
@@ -779,23 +768,6 @@ impl core::ops::Deref for DocNodeWithContext {
 
   fn deref(&self) -> &Self::Target {
     &self.inner
-  }
-}
-
-pub fn setup_highlighter(
-  show_line_numbers: bool,
-) -> comrak_adapters::HighlightAdapter {
-  comrak_adapters::HighlightAdapter {
-    #[cfg(feature = "syntect")]
-    syntax_set: syntect::dumps::from_uncompressed_data(include_bytes!(
-      "./default_newlines.packdump"
-    ))
-    .unwrap(),
-    #[cfg(feature = "syntect")]
-    theme_set: syntect::highlighting::ThemeSet::load_defaults(),
-    #[cfg(feature = "tree-sitter")]
-    language_cb: tree_sitter::tree_sitter_language_cb,
-    show_line_numbers,
   }
 }
